+++ docs/extras/modules/model_io/models/llms/integrations/aws_sagemaker.mdx
@@ -1,6 +1,6 @@
-# AWS SageMakerEndpointWithStream
+# AWS SageMakerEndpoint
 
-LangChain.js supports integration with AWS SageMaker-hosted endpoints also using streams. Check [Amazon SageMaker JumpStart](https://aws.amazon.com/sagemaker/jumpstart/) for a list of available models, and how to deploy your own.
+LangChain.js supports integration with AWS SageMaker-hosted endpoints. Check [Amazon SageMaker JumpStart](https://aws.amazon.com/sagemaker/jumpstart/) for a list of available models, and how to deploy your own.
 
 Here's an example:
