+++ langchain/.gitignore
@@ -148,9 +148,9 @@ llms/googlepalm.d.ts
 llms/fireworks.cjs
 llms/fireworks.js
 llms/fireworks.d.ts
-llms/sagemaker.cjs
-llms/sagemaker.js
-llms/sagemaker.d.ts
+llms/sagemaker_endpoint.cjs
+llms/sagemaker_endpoint.js
+llms/sagemaker_endpoint.d.ts
 llms/bedrock.cjs
 llms/bedrock.js
 llms/bedrock.d.ts
