+++ langchain/package.json
@@ -160,9 +160,9 @@
     "llms/fireworks.cjs",
     "llms/fireworks.js",
     "llms/fireworks.d.ts",
-    "llms/sagemaker.cjs",
-    "llms/sagemaker.js",
-    "llms/sagemaker.d.ts",
+    "llms/sagemaker_endpoint.cjs",
+    "llms/sagemaker_endpoint.js",
+    "llms/sagemaker_endpoint.d.ts",
     "llms/bedrock.cjs",
     "llms/bedrock.js",
     "llms/bedrock.d.ts",
@@ -1450,10 +1450,10 @@
       "import": "./llms/fireworks.js",
       "require": "./llms/fireworks.cjs"
     },
-    "./llms/sagemaker": {
-      "types": "./llms/sagemaker.d.ts",
-      "import": "./llms/sagemaker.js",
-      "require": "./llms/sagemaker.cjs"
+    "./llms/sagemaker_endpoint": {
+      "types": "./llms/sagemaker_endpoint.d.ts",
+      "import": "./llms/sagemaker_endpoint.js",
+      "require": "./llms/sagemaker_endpoint.cjs"
     },
     "./llms/bedrock": {
       "types": "./llms/bedrock.d.ts",
