+++ langchain/scripts/create-entrypoints.js
@@ -64,7 +64,7 @@ const entrypoints = {
   "llms/googlevertexai/web": "llms/googlevertexai/web",
   "llms/googlepalm": "llms/googlepalm",
   "llms/fireworks": "llms/fireworks",
-  "llms/sagemaker": "llms/sagemaker",
+  "llms/sagemaker_endpoint": "llms/sagemaker_endpoint",
   "llms/bedrock": "llms/bedrock",
   "llms/llama_cpp": "llms/llama_cpp",
   "llms/writer": "llms/writer",
@@ -310,7 +310,7 @@ const requiresOptionalDependency = [
   "llms/hf",
   "llms/raycast",
   "llms/replicate",
-  "llms/sagemaker",
+  "llms/sagemaker_endpoint",
   "llms/bedrock",
   "llms/llama_cpp",
   "llms/writer",
